feat: refactor type

+++ modules/sui_sdk/src/sui_transaction.h
@@ -8,8 +8,8 @@ class SuiTransaction
 {
 protected:
 public:
-	String signTransaction(String sender, String recipient, unsigned long long int amount);
-	String programmableTransactionAllowSponser(String sender_address, String recipient_address, unsigned long long int amount, String sponser_address);
+	String signTransaction(String sender, String recipient, uint64_t amount);
+	String programmableTransactionAllowSponser(String sender_address, String recipient_address, uint64_t amount, String sponser_address);
 	String requestTokensFromFaucet(String address_str);
 };
